backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.1.0
numpy>=1.25.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
supabase>=2.0.0
python-dotenv>=1.0.0
